rename all dropdown files to deprecated
