+++ src/docs/navigation.js
@@ -45,7 +45,7 @@ import MathSymbols from '../components/math-symbols/pages/math-symbols';
 import separators from '../components/separators/pages/separators';
 import spinners from '../components/spinner/pages/spinners';
 import SpinnerContainers from '../components/spinner-container/pages/spinner-containers';
-import dropdownsDeprecated from '../components/dropdowns/pages/dropdowns';
+import dropdownsDeprecated from '../components/dropdowns-deprecated/pages/dropdowns-deprecated';
 
 const navigation = [
   {
